array

+++ Car.cpp
@@ -4,5 +4,5 @@ using namespace std;
 Car::Car(int id) : Vehicle(id) {}
 
 int Car::getParkingDuration() {
-    return static_cast<int> (time(nullptr) - timeOfEntry)*0.9; //statics_cast<int> = convert to int
+    return (time(nullptr) - timeOfEntry)*0.9; //statics_cast<int> = convert to int
 }
